Move example helpers to module
